microservices/python/personalization-engine/requirements.txt
aiofiles>=23.2.1
boto3>=1.34.0
botocore>=1.34.0
faiss-cpu>=1.7.4
fastapi>=0.109.0
loguru>=0.7.2
numpy>=1.24.4
pydantic>=2.5.3
python-dotenv>=1.0.0
redis>=5.0.1
sentence-transformers>=2.2.2
uvicorn>=0.27.0
nats-py==2.6.0
tenacity==8.2.3
prometheus-client==0.17.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
pybreaker==1.0.1
-e ../common/nats_lib
